Make page and blog post editing experience consistent
When editing/creating a blog post or a page, the "save" button always redirects to the "edit" page, with a notification message at the top stating the new state of the page (draft, published, scheduled for publishing).

This makes the experience consistent with Wordpress blog editing, and also fixes a Pri1 bug about a 404 when creating a draft post.

This also fixes a bug where "publish later" wasn't working (strings are evil) for either blog of page.