Fix scheduled publishing of pages
Data model is now correctly updated when publishing, creating drafts, scheduling publishing, etc. There is no code left in the pages module which directly sets the CommonAspect dates. All is properly done through the CommonAspectHandler, the ContentManager and the PublishingTaskManager.

There is a remaining issue with the background scheduler which is unable to run scheduled tasks (exception thrown by autofac). This will be investigated later.