Make extension loading a 2 pass process
First pass is "probing": any loader who thinks it can load a
given extension returns a descriptor for it. The descriptor contains
information about how old the extension is.

The extension manager sorts the descriptor list return by probing all
loaders, and picks the most recent entry.

During the 2nd pass, each loader is called again, with the candidate extension
entry as argument. This allows each loader to be notified that, if they don't
load an extension, they might still need to cleanup some state related to
previous loading.

In essence, this is implementing a complete state machine of loaders and
extension "active/inactive" states/transitions.