Fix SpecFlow test
When a new module is installed in the host app, explicitly call
"ReloadExtensions" on the host so that the new module is activated.

We need this because some SpecFlow binding talk directly to the host through
remoting, whereas in the real web scenarios, all comunication is done through
"BeingRequest" which is the guy taking care of reloading modules when
they have changed on disk.