Fix feature enable from command line
When enabling features from the command line, we need
a way to signal the WebHost app that the OrchardHost needs
to re-compute the list of activated features.

We enable this by touching and monitoring a local file in "App_Data".
The file is touched whenever a feature is enabled or disabled,
and is monitored by the OrchardHost to watch for a re-initialization.