Updating the RoutableHomePageProvider to use the relevant display action for the given content item.
This lets unique display actions, like Blog/Item, do their own thing. In the case of the blog, its posts are once again showing up on the home page.

The routing problem (where the home page'd item still is seen as having its own path) still needs to be fixed so the item on the home page can't be hit at different URLs (and so something like paging works on the home page)

work item: 16720