Improve response time for first time users
On the first time installation of Orchard, the user gets to the setup screen, which
will take a while to finish (user entering data and the setup process itself).
We use this opportunity to start a background task to "pre-compile" all the known
views in the app folder, so that the application is more reponsive when the user
hits the homepage and admin screens for the first time.