SpecFlow test for dynamic compilation
The SpecFlow test for dynamic compilation wasn't actually trigering
dynamic compilation for all modules, because some of the were copied
to the "~/bin" assembly of the hosted ASP.NET application used for
SpecFlow testing.

This now works as expected, and we have 3 tests for this: one with
dynamic compilation enabled, one with dynamic comilation disabled,
and one with forced dynamic compilation (other loaders disabled).