Adding back a DisplayType shape display alternate for the "Content" shape.
Also tweaked the order of alternates to minimize the need for additional templates when any non-DisplayType contained alternate is used.

In other words. If there is a Content-BlogPost.Summary template and a Content-42 template is used to override the detail display of a blog post with an id of 42 - then a Content-42.Summary is not necessary.