Fix 1.0 to 1.1 migration issue
SiteSettings are used very early and often in Orchard, so we
can't rely on having the user be able to run migration for
SiteSettings. The workaround for 1.1 is to move the changes
we did to SiteSettings into a new SiteSettings2 Part/Record.

With this change, database compatiblity from 1.0 and 1.1 seems
to be in a good working state on a real blog site.