#17804: Ensure views are recompiled more aggressively
When a module changes, it's loader can change, which means the list
of dependencies can change to. Instead of trying to track the list
of dependencies, we now simply add a dependency on a file on disk
which describe the module compilation state.

As an aside, this also improves startup performance when
there were no changes to a site, since asp.net now only has
to check a single file instead of the list of dependencies.