Refactoring OwnerEditor and DateEditor breakout
Splitting OwnerEditor and DateEditor settings apart. Breaking off any knowledge
in the CommonPart of the OwnerEditor and DateEditor.

Changed the settings to appear as TypePart definition instead of Type.

Based the ViewModel on Shape instead of EditorTemplates.

The goal of these changes is to illustrate how a 3rd party module would
accomplish the same kind of ContentPart extension w/out changing core code.