#17728: Workaround asp.net bug loading incorrect version of System.
Some of our assemblies in "~/bin" reference "System, Version=2.0.0.0".
During the first time site compilation in the command line scenario,
asp.net will sometimes decide to load that version of the assembly
from disk (from the GAC), even though the application is configured
to run on .NET 4.0. The workaround is to remove a couple of web.config
elements, so that ASP.NET will load the correct version of those
assemblies at startup.

Work Item: 17728